Add ways to mute misbehaving Graph nodes
The RCA framework supports arbitrary code in the operate() methods for Symptom and RCA nodes.
Therefore, there can be cases where this code can throw exceptions and cause. This patch
introduces a way to mute these nodes. Muting a node means that the node's operate will not be
called on subsequent runs and a call to their getFlowUnit will return an empty flow unit.

The bais for muting nodes:
- Number of exceptions thrown in the operate method per node.

This patch introduces only one criterion to mute a node. But the framework is extensible.
We should be able to write muters based on any measurement we capture.

Tests:
Unit tests are added for this path and covers the code added here.