Exception handling logic change for metrics gather method (#91)
Metrics gather(), whihc calls the MetricsDB to get the metrics can throw exception if
the queried metric does not exist. We were handling the exception earlier than we should
and hence the gather was not returning an empty flow unit. We were alos missing on logging
the exception at teh right place and unnecessarily logging the entire stack trace which
is not required as this is expected.