bug: handling the case when thread id no longer exists.
If the threads in the thread pool gets shuffled for whatever reason, then the threadinfo for a thread that existed in the recent past but no longer does. In that case the ThreadInfo for that corresponding id will be null. If a null pointer exception is thrown because of this, then the request might error out on the Elasticsearch side. The change here catches the exception and skips over such ids.

Testing: