Fix performance-analyzer-agent configFilePath (#268)
We have a class Util.java which we use to look up the location of the
performance-analyzer.properties file. The Util class prepends a prefix
to the Java property we pass in as -DconfigFilePath. This prefix
resulted in a FileNotFoundException error when attempting to locate the
config file inside our Docker environment.

This commit fixes the issue by updating the -DconfigFilePath variable in
the performance-analyzer-agent script file. This is a stopgap measure.
We should strive to refactor the Util class so that it's configurable
and working across all supported platforms.