Rest mutual auth fix (#279)
* Add required mutual auth to gRPC Server/Client

Previously we had 1 sided TLS on the server side. Data between the
client and server was send over an encrypted channel, but any client
could make requests to the server.

This commit changes the behavior so that only clients with the matching
certificates can make requests to the server when TLS is enabled. This
commit does NOT add support for installing a trust manager. That must be
added in the future.

* Add full mutual auth to gRPC client/server and augment tests

* Implement mutual auth for the REST endpoints

This commit makes the PerformanceAnalyzerWebServer authenticate clients
if the user specifies a certificate authority. It also properly sets up
the server's identity, so that any clients can authenticate the server.

* Fix merge issue with WireHopperTest

* Fixing up PerformanceAnalyzerWebServerTest

* Modify gradle.yml for testing

* Remove info log flooding from testing gradle.yml