Integration test framework to test RCAs and decision Makers  (#301)
* Integration test framework to test RCAs and decision Makers

This PR introduces a framework that can be used to write tests concerning
RCA graph and all the components that leverage it(RCAs, Deciders etc.).
Currently it supports adding custom metrics, running tests against
single node cluster, multi-node clusters with dedicated masters and
the ones without dedicated masters. It provies a way to query the
SQLite datastore file on a particular host or hit the rest endpoint
on a particular node. Currently, the nodes of the clsuters work over
http only.

* Fixed a bug where the database was being corrupted because of multiple scheduler initializations.

* Fix the POC graph

* Removing unused methods

* Name the validation checks that failed

* Addressing PR comments