Populate default workload type and cache priority for the decider to base default actions (#340)
Take workload type / cache priority order as an input from the customer so as to default decider actions based on these inputs. Also have a generic decider config which would be populated whenever any decider node is added to the RCA graph.

These changes take the workload type and cache priorities as an input from the customer to base the actions taken on the cluster. Default values are hardcoded in the code in case the read from config is not possible due to any circumstances. Values present in the config file are given priority over any such value.

Co-authored-by: Aditya Jindal <aditjind@amazon.com>