Add the BucketCalculator class and a few basic impls (#391)
* Add the BucketCalculator class and a few basic impls

- The BucketCalculator class can be used to compute a "UsageBucket" for
a particular resource. E.g. if CPU utilization is at 99% utilization
a BucketCalculator may say that CPU is unhealthy. At the other extreme,
if CPU utilization is at 1%, a BucketCalculator may say that CPU is
underutilized (which means that we may be able to remove a node from our
cluster, for example).

* Fixup unit tests

* Change SubscribeResponseHandler#onError to log a warning

* Address PR Comments

- Remove JvmBucketCalculator
- Rename StaticBucketCalculator to BasicBucketCalculator
- Update UsageBucket enum comments

* Clean up PR

* Update logic to validate bucket ranges

* Fix checkstyle issues