Wrongly updating all three variants of rca.conf with the current one.
When we update the rca.conf with the new set of muted rcas/deciders/actions, the idea
is to pick each of rca.conf, rca_master.conf and rca_idle_master.conf
and update each of them with the latest set of muted *, keeping everything
else constant. The code before this would have made all files exactly alike.