Changing metricsDB provider while executing an integTest (#523)
* Changing metricsDB provider while executing an integTest

Today the gauntlet only let's you provide the metricsDB table at the top of a test class
or at the top of a test method. But sometimes during the test test execution, one might want
to change the instance of metricsDB. This is made possible by this patch. In order to do that,

You create a new inner class (doesn't have to be inner) with an appropriate name. Add, the metrics
details as you would if you were to do it as part of the main test class itself.

  @AMetric(name = CPU_Utilization.class,
      dimensionNames = {SHARDID_VALUE, INDEX_NAME_VALUE, OPERATION_VALUE, SHARD_ROLE_VALUE},
      tables = {
          @ATable(hostTag = HostTag.DATA_0,
              tuple = {
                  @ATuple(dimensionValues = {"0", "logs", "bulk", "p"},
                      sum = 0.0, avg = 0.0, min = 0.0, max = 111.0),
                  @ATuple(dimensionValues = {"1", "logs", "bulk", "r"},
                      sum = 0.0, avg = 0.0, min = 0.0, max = 88.0),
                  @ATuple(dimensionValues = {"2", "logs", "bulk", "p"},
                      sum = 0.0, avg = 0.0, min = 0.0, max = 11.0)
              }
          ),
      }
  )
  public static class MetricsAfterEventX {

  }

then from your test method, you can do something like this:

  @Test
  @AExpect(
      what = AExpect.Type.REST_API,
      on = HostTag.DATA_0,
      validator = UpdatedMetricsPocValidator.class,
      forRca = RcaItPocDedicated.SimpleAnalysisGraphForDedicated.ClusterRca.class)
  public void testMetricsUpdated() throws Exception {
    // Initial logic, uses the Class default metricsDB choice.
    // ...

    api.updateMetrics(MetricsAfterEventX.class);

   // Other test logic follows.
  }

* remove the debug log

* added some more comments