Update the metricsDB instance for the gauntlet tests without a scheduler restart (#528)
* Update the metricsDB instance for the gauntlet tests without a scheduler restart.

Today the metricsDB instance serving the graph node can be updated but for the
update to take effect, it also restarts the Scheduler. The scheduler stores the
states of the RCA nodes in memory and therefore restarting it cases all transient
states to be lost. This change tries to make the chnage without causing a Scheduler
restart. In order to do this, we add a check in the SchedulerTask. Before every
run invokation, it checks if someone has asked for the metricsDB to be updated
if so, then it goes ahead and updates the reference for all Tasklets. This is only
for test purposes and should not happen in prod.

* Removed unused imports